// frontend/src/api/model/createLabel.ts
/**
 * Efficient Labeling of Time Series
 * No description provided
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Project - Create request object name: The name of this InlineObject1 [Optional]. file: The file of this InlineObject1 [Optional].
 */
export interface CreateLabel { 
    label_class?: string;
    start?: number;
    end?: number;
    sample?: number;
    dimensions?: Array<number>;
}