// frontend/src/api/model/prediction.ts
/**
 * Efficient Labeling of Time Series
 * No description provided
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { SamplePrediction } from './samplePrediction';
import { PredictionAlgorithm } from './predictionAlgorithm';


/**
 * Prediction Object
 */
export interface Prediction { 
    model?: PredictionAlgorithm;
    params?: object;
    sample_predictions?: Array<SamplePrediction>;
}