// frontend/src/api/model/samplePrediction.ts
/**
 * Efficient Labeling of Time Series
 * No description provided
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { PredictionWindow } from './predictionWindow';


export interface SamplePrediction { 
    sample?: number;
    prediction?: Array<PredictionWindow>;
}