// frontend/src/api/model/timeSeries.ts
/**
 * Efficient Labeling of Time Series
 * No description provided
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * TimeSeries - a model defined in OpenAPI  id: The id of this TimeSeries [Optional]. data: The data of this TimeSeries [Optional]. timestamps: The timestamps of this TimeSeries [Optional].
 */
export interface TimeSeries { 
    id?: number;
    data?: Array<number>;
    data_max?: Array<number>;
    data_min?: Array<number>;
    timestamps?: Array<number>;
}